--- plugins/dev/1778.go
@@ -1,44 +1,0 @@
-package dev
-
-import (
-	"megaCrawler/crawlers"
-	"megaCrawler/extractors"
-
-	"github.com/gocolly/colly/v2"
-)
-
-func init() {
-	engine := crawlers.Register("1778", "Evening Gazette", "https://www.gazettelive.co.uk/")
-
-	engine.SetStartingURLs([]string{
-		"https://www.gazettelive.co.uk/news/teesside-news/",
-		"https://www.gazettelive.co.uk/all-about/hartlepool",
-		"https://www.gazettelive.co.uk/all-about/darlington",
-		"https://www.gazettelive.co.uk/news/uk-world-news/",
-	})
-
-	extractorConfig := extractors.Config{
-		Author:       true,
-		Image:        true,
-		Language:     true,
-		PublishDate:  true,
-		Tags:         true,
-		Text:         false,
-		Title:        true,
-		TextLanguage: "",
-	}
-
-	extractorConfig.Apply(engine)
-
-	engine.OnHTML(".headline", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Visit(element.Attr("href"), crawlers.News)
-	})
-
-	engine.OnHTML(".article-body", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Content += element.Text
-	})
-
-	engine.OnHTML(".pagi-next > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Visit(element.Attr("href"), crawlers.Index)
-	})
-}
