+++ plugins/storage/1780.go
@@ -1,0 +1,42 @@
+package storage
+
+import (
+	"strings"
+
+	"megaCrawler/crawlers"
+	"megaCrawler/extractors"
+
+	"github.com/gocolly/colly/v2"
+)
+
+func init() {
+	engine := crawlers.Register("1780", "Watts Up With That?", "https://wattsupwiththat.com/")
+
+	engine.SetStartingURLs([]string{"https://wattsupwiththat.com/sitemap-index-1.xml"})
+
+	extractorConfig := extractors.Config{
+		Author:       true,
+		Image:        true,
+		Language:     true,
+		PublishDate:  true,
+		Tags:         true,
+		Text:         false,
+		Title:        true,
+		TextLanguage: "",
+	}
+
+	extractorConfig.Apply(engine)
+
+	engine.OnXML("//loc", func(element *colly.XMLElement, ctx *crawlers.Context) {
+		switch {
+		case strings.Contains(element.Text, "sitemap"):
+			engine.Visit(element.Text, crawlers.Index)
+		default:
+			engine.Visit(element.Text, crawlers.News)
+		}
+	})
+
+	engine.OnHTML(".entry-content.magazinenp-parts-item", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Content += element.Text
+	})
+}
